requirements.txt
streamlit
faiss-cpu
pandas
openpyxl
beautifulsoup4
transformers
sentence-transformers
scikit-learn
numpy
